refactor: promote client data to shared and adopt Oni mascot
AI루틴 탭(다음 커밋)이 고객 선택·식단 요약에 고객 데이터를 쓰는데,
STRUCTURE.md의 no-cross-feature-import 규칙에 따라 공유 레이어로 승격.

- TrainerClient 엔티티 → shared/models, ClientRepository(+providers)
  → shared/services (여러 feature 공유 데이터는 shared로 — STRUCTURE §2.4)
- OniAvatar(온케어 마스코트) shared/widgets 추가 — 사용자 앱 리디자인
  (feature/188 figma kit)과 동일한 AI 아이덴티티, 트레이너 앱은 사용자
  앱 패키지를 import하지 않으므로 재구현. AI 요약 카드의 auto_awesome
  아이콘을 Oni로 교체(디자인 일관성)

flutter analyze: no issues
