+++ frontend/flutter_trainer/lib/shared/models/trainer_client.dart
